refactor: Move constants to utils
