+++ src/components/LinkListItem/index.js
@@ -8,7 +8,7 @@ export const StyledLink = styled.a`
   background-image: linear-gradient(150deg, #c3cfe2 0%, #f5f7fa 100%);
   background-image: linear-gradient(-20deg, #b721ff 0%, #21d4fd 100%);
   background-size: 100% 0.3rem;
-  background-position: 0 110%;
+  background-position: 0 115%;
   transition: background-size 100ms cubic-bezier(0.5, 0, 0.1, 1);
   color: currentColor;
